refactor(testutil): move starwars utils to own file
